@frederickfogerty 2019-07-16
Thanks for the PR @baldurh. This all looks good to me, now we just have to wait for the folks over at imgix to look at this and merge this in.